Enable multiple input modalities
Added the option to indicate in self.input_modalities which modalities the method takes as input (can be one or multiple). Overloaded _load_input_image to load in a list of images to pass to predict. Needed to skip UniquePathIndicesValidator as this errors when multiple images (modalities) are present in the input folder per subject (which is the case when running with multiple modalities as input)